update sistem login

+++ award/views.py
@@ -14,40 +14,44 @@ def get_delete_update_award(request, pk):
    
     try:
         Award = award.objects.get(pk=pk)
-        try:
-            token = request.META.get('HTTP_AUTHORIZATION','')
-            get_token = Register.objects.get(token = token) 
-            if (get_token.id == Award.id):
-                if request.method == 'GET':
-                    serializer = AwardSerializer(Award)
-                    return Response(serializer.data)
+        if (registrations.token == 'xxx'):
+            response = {'status':'LOGIN FIRST, YOU MUST...'}
+            return Response(response, status=status.HTTP_401_UNAUTHORIZED)
+        else:    
+            try:
+                token = request.META.get('HTTP_AUTHORIZATION','')
+                get_token = Register.objects.get(token = token) 
+                if (get_token.id == Award.id):
+                    if request.method == 'GET':
+                        serializer = AwardSerializer(Award)
+                        return Response(serializer.data)
 
-                elif request.method == 'DELETE':                
-                    Award.delete()
+                    elif request.method == 'DELETE':                
+                        Award.delete()
+                        content = {
+                            'status' : 'NO CONTENT'
+                        }
+                        return Response(content, status=status.HTTP_202_NO_CONTENT)
+                        
+                         
+                    elif request.method == 'PUT':
+
+                            serializer = AwardSerializer(Award, data=request.data)
+                            if serializer.is_valid():
+                                serializer.save()
+                                return Response(serializer.data, status=status.HTTP_201_CREATED)
+                            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                else:
                     content = {
-                        'status' : 'NO CONTENT'
+                    'status': 'UNAUTHORIZED'
                     }
-                    return Response(content, status=status.HTTP_202_NO_CONTENT)
+                    return Response(content, status=status.HTTP_401_UNAUTHORIZED)
                     
-                     
-                elif request.method == 'PUT':
-
-                        serializer = AwardSerializer(Award, data=request.data)
-                        if serializer.is_valid():
-                            serializer.save()
-                            return Response(serializer.data, status=status.HTTP_201_CREATED)
-                        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-            else:
+            except Register.DoesNotExist:
                 content = {
-                'status': 'UNAUTHORIZED'
+                    'status': 'UNAUTHORIZED'
                 }
                 return Response(content, status=status.HTTP_401_UNAUTHORIZED)
-                
-        except Register.DoesNotExist:
-            content = {
-                'status': 'UNAUTHORIZED'
-            }
-            return Response(content, status=status.HTTP_401_UNAUTHORIZED)
 
     except award.DoesNotExist:
         content = {
