+++ experiences/views.py
@@ -12,41 +12,45 @@ from registrations.models import Register
 def get_delete_update_experiences(request, pk):
     try:
         Experiences = pengalaman.objects.get(pk=pk)
-        try:
-            token = request.META.get('HTTP_AUTHORIZATION','')
-            get_token = Register.objects.get(token = token)
-            if (get_token.id == Experiences.id):
+        if (registrations.token == 'xxx'):
+            response = {'status':'LOGIN FIRST, YOU MUST...'}
+            return Response(response, status=status.HTTP_401_UNAUTHORIZED)
+        else:    
+            try:
+                token = request.META.get('HTTP_AUTHORIZATION','')
+                get_token = Register.objects.get(token = token)
+                if (get_token.id == Experiences.id):
 
-                if request.method == 'GET':
-                    serializer = ExperiencesSerializer(Experiences)
-                    return Response(serializer.data)
+                    if request.method == 'GET':
+                        serializer = ExperiencesSerializer(Experiences)
+                        return Response(serializer.data)
 
-                elif request.method == 'DELETE':
-                
-                    Experiences.delete()
-                    content = {
-                        'status' : 'NO CONTENT'
-                    }
-                    return Response(content, status=status.HTTP_202_NO_CONTENT)
+                    elif request.method == 'DELETE':
                     
-                elif request.method == 'PUT':                
-                    serializer = ExperiencesSerializer(Experiences, data=request.data)
-                    if serializer.is_valid():
-                        serializer.save()
-                        return Response(serializer.data, status=status.HTTP_201_CREATED)
-                    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                        Experiences.delete()
+                        content = {
+                            'status' : 'NO CONTENT'
+                        }
+                        return Response(content, status=status.HTTP_202_NO_CONTENT)
+                        
+                    elif request.method == 'PUT':                
+                        serializer = ExperiencesSerializer(Experiences, data=request.data)
+                        if serializer.is_valid():
+                            serializer.save()
+                            return Response(serializer.data, status=status.HTTP_201_CREATED)
+                        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-            else:
+                else:
+                    content = {
+                    'status': 'UNAUTHORIZED'
+                    }
+                    return Response(content, status=status.HTTP_401_UNAUTHORIZED)
+                
+            except Register.DoesNotExist:
                 content = {
-                'status': 'UNAUTHORIZED'
+                    'status': 'UNAUTHORIZED'
                 }
                 return Response(content, status=status.HTTP_401_UNAUTHORIZED)
-            
-        except Register.DoesNotExist:
-            content = {
-                'status': 'UNAUTHORIZED'
-            }
-            return Response(content, status=status.HTTP_401_UNAUTHORIZED)
     except pengalaman.DoesNotExist:
         content = {
             'status': 'Not Found'
